Remove redundant base class.
Reintegrate CommentAdapterBase into CommitNoteAdapter, and make it use
the timeline comment row layout.
