--- app/src/main/res/layout/row_gravatar_comment.xml
@@ -1,117 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:background="?attr/selectableItemBackground"
-    android:clipToPadding="false"
-    android:minHeight="?attr/listPreferredItemHeight"
-    android:padding="@dimen/content_padding">
-
-    <FrameLayout
-        android:id="@+id/fl_avatar_container"
-        android:layout_width="32dp"
-        android:layout_height="32dp"
-        android:layout_marginRight="8dp">
-
-        <com.gh4a.widget.FixedSizeImageView
-            android:id="@+id/iv_gravatar"
-            android:layout_width="24dp"
-            android:layout_height="24dp"
-            android:background="?attr/selectableItemBackgroundBorderless"
-            tools:src="@drawable/default_avatar" />
-
-        <com.gh4a.widget.FixedSizeImageView
-            android:id="@+id/iv_event_icon"
-            android:layout_width="16dp"
-            android:layout_height="16dp"
-            android:layout_gravity="end|bottom"
-            android:visibility="gone"
-            tools:src="@drawable/issue_event_closed"
-            tools:visibility="visible" />
-    </FrameLayout>
-
-    <ImageView
-        android:id="@+id/iv_menu"
-        style="@style/SelectableBorderlessItem"
-        android:layout_width="24dp"
-        android:layout_height="24dp"
-        android:layout_alignParentRight="true"
-        android:layout_marginRight="-8dp"
-        android:scaleType="center"
-        android:src="?overflowIcon" />
-
-    <com.gh4a.widget.StyleableTextView
-        android:id="@+id/tv_extra"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:layout_toRightOf="@id/fl_avatar_container"
-        android:textAppearance="?android:attr/textAppearanceSmall"
-        android:textColor="?android:attr/textColorPrimary"
-        tools:text="username" />
-
-    <com.gh4a.widget.StyleableTextView
-        android:id="@+id/tv_timestamp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignBaseline="@id/tv_extra"
-        android:layout_alignWithParentIfMissing="true"
-        android:layout_marginLeft="16dp"
-        android:layout_toLeftOf="@id/iv_menu"
-        android:layout_toRightOf="@id/tv_extra"
-        android:textAppearance="@style/TextAppearance.VerySmall"
-        tools:text="5 days ago" />
-
-    <com.gh4a.widget.StyleableTextView
-        android:id="@+id/tv_edit_timestamp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@id/tv_timestamp"
-        android:layout_marginLeft="16dp"
-        android:layout_toLeftOf="@id/iv_menu"
-        android:layout_toRightOf="@id/tv_extra"
-        android:drawableLeft="?attr/commentEditIcon"
-        android:drawablePadding="2dp"
-        android:textAppearance="@style/TextAppearance.VerySmall"
-        tools:text="5 days ago" />
-
-    <com.gh4a.widget.StyleableTextView
-        android:id="@+id/tv_file"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignWithParentIfMissing="true"
-        android:layout_below="@id/tv_edit_timestamp"
-        android:layout_toLeftOf="@id/iv_menu"
-        android:layout_toRightOf="@id/fl_avatar_container"
-        android:textAppearance="?android:attr/textAppearanceSmall"
-        android:visibility="gone"
-        tools:text="Comment on filename.txt"
-        tools:visibility="visible" />
-
-    <com.gh4a.widget.StyleableTextView
-        android:id="@+id/tv_desc"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignWithParentIfMissing="true"
-        android:layout_below="@id/tv_file"
-        android:layout_marginTop="2dp"
-        android:layout_toLeftOf="@id/iv_menu"
-        android:layout_toRightOf="@id/fl_avatar_container"
-        android:textAppearance="?android:attr/textAppearanceSmall"
-        android:textColor="?android:attr/textColorPrimary"
-        android:textIsSelectable="true"
-        tools:text="Comment text" />
-
-    <com.gh4a.widget.ReactionBar
-        android:id="@+id/reactions"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignWithParentIfMissing="true"
-        android:layout_below="@id/tv_desc"
-        android:layout_marginTop="2dp"
-        android:layout_toLeftOf="@id/iv_menu"
-        android:layout_toRightOf="@id/fl_avatar_container" />
-
-</RelativeLayout>
